minor data fix

+++ Turnero.DAL/Models/GrowthChart.cs
@@ -2,13 +2,13 @@
 
 public class GrowthChart : BaseEntity
 {
-    public int Age { get; set; }
+    public double Age { get; set; }
     public string? Time { get; set; }
     public double Weight { get; set; }
     public string? WPerc { get; set; }
-    public int Height { get; set; }
+    public double Height { get; set; }
     public string? HPerc { get; set; }
-    public int HeadCircumference { get; set; }
+    public double HeadCircumference { get; set; }
     public string? HCPerc { get; set; }
     public double Bmi { get; set; }
     public Patient? Patient { get; set; }
